// frontend/src/app/api-client/model/tokenPair.ts
/**
 * Consola de Operaciones Transaccionales API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TokenPair { 
    accessToken: string;
    refreshToken: string;
    tokenType: string;
    /**
     * Vida del access token en segundos.
     */
    expiresIn: number;
}